docs, unregistering should unregister overrides, tests
